pyproject.toml
[project]
name = "aiagent14-job-agents-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "celery>=5.6.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "google-cloud-speech>=2.27.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.80",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.3",
    "pdfkit>=1.0.0",
    "pillow>=12.0.0",
    "playwright>=1.48.0",
    "matplotlib>=3.9.0",
    "openpyxl>=3.1.5",
    "jinja2>=3.1.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pydub>=0.25.1",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "redis>=5.0.0",
    "reportlab>=4.4.5",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "tiktoken>=0.8.0",
    "uvicorn>=0.35.0",
    "websockets>=12.0",
    "seaborn>=0.13.2",
    "plotly>=6.5.0",
    "pymysql>=1.1.2",
    "cryptography>=46.0.3",
    "audioop-lts>=0.2.2",
    "google-genai>=0.3.0",
]
